feat: integrate useExplorerResults hook and enhance ExplorerBox and ExplorerDetailsModal components for improved data handling
