+++ DashAI/front/src/components/notebooks/explorer/tabs/Results.jsx
@@ -1,231 +1,17 @@
-import React, { useEffect, useState } from "react";
-
-import { Box, CircularProgress, Tooltip, Typography } from "@mui/material";
-import { getExplorerResults } from "../../../../api/explorer";
+import React from "react";
+import { Box, CircularProgress } from "@mui/material";
+import { visualizersKeys } from "../useExplorerResults";
 import ImageVisualizer from "../visualizations/ImageVisualizer";
 import PlotlyJsonVisualizer from "../visualizations/PlotlyJsonVisualizer";
 import TabularVisualizer from "../visualizations/TabularVisualizer";
-import PlotLayoutForm from "../plotLayout/PlotLayoutForm";
-import { updateExplorerResults } from "../../../../api/explorer";
-import { useSnackbar } from "notistack";
-
-/**
- * NullCell component to render null values in the tabular visualizer
- * @param {Object} props
- */
-function NullCell({}) {
-  const [hover, setHover] = useState(false);
-  return (
-    <Box
-      sx={{
-        width: "100%",
-        height: "100%",
-        display: "flex",
-        alignItems: "center",
-      }}
-      onMouseEnter={() => setHover(true)}
-      onMouseLeave={() => setHover(false)}
-    >
-      <Typography variant="body2" color="text.disabled">
-        {hover ? "None" : "-"}
-      </Typography>
-    </Box>
-  );
-}
-
-const visualizers = {
-  tabular: TabularVisualizer,
-  plotly_json: PlotlyJsonVisualizer,
-  image_base64: ImageVisualizer,
-  image_url: ImageVisualizer,
-};
-const visualizersKeys = {
-  tabular: "tabular",
-  plotly_json: "plotly_json",
-  image_base64: "image_base64",
-  image_url: "image_url",
-};
-
-const ORIENTATIONS = {
-  dict: "dict",
-  records: "records",
-};
-
-/**
- * Get the data from the orientation given. This function is used to transform the data
- * from the explorer results to the format required by the tabular visualizer.
- * @param {Object} data The data from the explorer results
- * @param {String} orientation The orientation of the data
- */
-const getDataFromOrientation = (data, orientation) => {
-  let res = {
-    columns: [],
-    rows: [],
-  };
-
-  if (orientation === ORIENTATIONS.records) {
-    throw new Error(`orientation ${orientation} not supported`);
-  }
-
-  if (orientation === ORIENTATIONS.dict) {
-    // ‘dict’ (default) : dict like {column -> {index -> value}}
-    // Get the columns
-    const columns = Object.keys(data);
-    res.columns = [
-      {
-        field: "id",
-        headerName: "Index",
-        renderCell: (params) => {
-          return (
-            <Tooltip title={params.value} arrow>
-              <Typography variant="body2">{params.value}</Typography>
-            </Tooltip>
-          );
-        },
-      },
-      ...columns.map((column) => {
-        return {
-          field: column,
-          headerName: column,
-          renderCell: (params) => {
-            if (params.value === null) {
-              return <NullCell />;
-            } else if (typeof params.value === "object") {
-              const tooltip = JSON.stringify(params.value);
-              return (
-                <Tooltip title={tooltip} arrow>
-                  <Typography variant="body2" color="text.secondary">
-                    {JSON.stringify(params.value)}
-                  </Typography>
-                </Tooltip>
-              );
-            } else if (
-              params.value !== "" &&
-              !isNaN(params.value) &&
-              !Number.isInteger(params.value)
-            ) {
-              const tooltip = params.value;
-              const display = parseFloat(params.value).toFixed(2);
-              return (
-                <Tooltip title={tooltip} arrow>
-                  <Typography variant="body2">{display}</Typography>
-                </Tooltip>
-              );
-            }
-            const tooltip = params.value;
-            return (
-              <Tooltip title={tooltip} arrow>
-                <Typography variant="body2">{params.value}</Typography>
-              </Tooltip>
-            );
-          },
-        };
-      }),
-    ];
-
-    // Get the rows
-    const rows = [];
-    const indexes = Object.keys(data[columns[0]]);
-    indexes.forEach((index) => {
-      const row = {
-        id: index,
-      };
-      columns.forEach((column) => {
-        row[column] = data[column][index];
-      });
-      rows.push(row);
-    });
-    res.rows = rows;
-  }
-
-  return res;
-};
 
 /**
  * Results component to render the results of the exploration
  * @param {Object} props
  * @param {Number} props.id The id of the exploration
  * @param {Boolean} props.minimalist Whether to render in minimalist mode with fixed dimensions
- * @param {Number} props.height The fixed height for minimalist mode (default: 300)
  */
-function Results({ id, minimalist = false }) {
-  const [loading, setLoading] = useState(false);
-  const [dataType, setDataType] = useState(null);
-  const [data, setData] = useState(null);
-  const { enqueueSnackbar } = useSnackbar();
-
-  const handleSaveChangesLayout = async () => {
-    try {
-      await updateExplorerResults(id, data);
-      enqueueSnackbar("Explorer results updated successfully", {
-        variant: "success",
-      });
-    } catch (error) {
-      console.error("Failed to update explorer results:", error);
-      enqueueSnackbar("Failed to update explorer results", {
-        variant: "error",
-      });
-    }
-  };
-
-  const fetchExplorerResults = async () => {
-    if (!id) return;
-    setLoading(true);
-    getExplorerResults(id)
-      .then((results) => {
-        if (!results?.type) {
-          throw new Error("No result type specified in the response");
-        }
-
-        // Check if there is an appropriate visualizer
-        if (!Object.keys(visualizers).includes(results.type)) {
-          throw new Error(`No visualizer found for type: ${results.type}`);
-        }
-        setDataType(results.type);
-
-        if (results.type === visualizersKeys.tabular) {
-          // Get the data from the orientation
-          const data = getDataFromOrientation(
-            results.data,
-            results.config.orient,
-          );
-          setData({
-            columns: data.columns.map((column) => {
-              return {
-                ...column,
-                // flex: 1,
-              };
-            }),
-            rows: data.rows,
-          });
-        }
-
-        if (results.type === visualizersKeys.plotly_json) {
-          setData(JSON.parse(results.data));
-        }
-
-        if (results.type === visualizersKeys.image_base64) {
-          setData(results.data);
-        }
-
-        if (results.type === visualizersKeys.image_url) {
-          setData(results.data);
-        }
-      })
-      .catch((error) => {
-        console.error(error);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
-
-  // Fetch the results data on mount
-  useEffect(() => {
-    if (!id) return;
-    fetchExplorerResults();
-  }, [id]);
-
+function Results({ id, minimalist = false, loading, data, dataType }) {
   if (!id) return null;
 
   const containerStyles = minimalist
@@ -275,25 +61,6 @@ function Results({ id, minimalist = false }) {
 
       {!loading && dataType === visualizersKeys.image_url && (
         <ImageVisualizer data={data} minimalist={minimalist} />
-      )}
-      {!minimalist && !loading && dataType === visualizersKeys.plotly_json && (
-        <PlotLayoutForm
-          data={data.data}
-          setData={(newData) => {
-            setData((prevData) => ({
-              ...prevData,
-              data: newData,
-            }));
-          }}
-          layout={data.layout}
-          setLayout={(newLayout) => {
-            setData((prevData) => ({
-              ...prevData,
-              layout: newLayout,
-            }));
-          }}
-          onSave={handleSaveChangesLayout}
-        />
       )}
     </Box>
   );
